Python Predictive Data Analytics
For this assignment, I decided to use the Sci-kit in order to develop a classifier that uses data to predict the outcome of a Bank marketing campaign. Using this library meant that I had to convert all descriptive features into a number representation. In order to do this, I used the binary one-hot encoding scheme. The DictVectorizer object was used to implement the binary one-hot encoding scheme. I also had to separate the target feature from the descriptive features  into different data structures. The target feature in this case was the predicted outcome. 
The next step was to actually train the decision tree. I used entropy based information gain in order to achieve this. I decided to use the entropy method as it seemed the most suitable with the dataset given. Information gain was retrieved through the decrease in entropy after the dataset was split based on an attribute. Information gain also allowed me to check which attribute is the most useful for splitting the dataset.
I also had to drop the question mark column of the queries file in order to make the dataset usable for the entropy based classifier. This was relatively easy to implement with the drop function in pandas. Once I ran each prediction, I stored them in a dataframe. I then dropped the columns which did not need to be displayed. Finally, I wrote the prediction outcome along with the test ID to the solutions file.